+++ src/Item.java
@@ -11,8 +11,8 @@ import IDs.AuctionHouseID;
 public class Item
 {
     public AuctionHouseID AUCTION_HOUSE_ID;      //IDs.ID of the AuctionHouse that holds this item
-    public final float MINIMUM_BID;         //Minimum bid to start at
-    private float currentBid;               //CurrentBid on this item--the highest of all the bids. Will be 0 until first bid.
+    public final Double MINIMUM_BID;         //Minimum bid to start at
+    private Double currentBid;               //CurrentBid on this item--the highest of all the bids. Will be 0 until first bid.
     public final String ITEM_NAME;          //Name of the item
     private String currentHighestBidderID;    //biddingID of Agent who holds the currentBid, which is winning. Will be null until first bid.
     private String imgPath;
@@ -20,18 +20,18 @@ public class Item
     private static int staticIDCounter = 1;
 
     
-    public Item(String itemName, String imgPath, float minimumBid)
+    public Item(String itemName, String imgPath, Double minimumBid)
     {
         ITEM_NAME = itemName;
         this.imgPath = imgPath;
         MINIMUM_BID = minimumBid;
-        currentBid = 0;
+        currentBid = 0.0;
         currentHighestBidderID = null;
         ITEM_ID = staticIDCounter*42 +""; //actually make it look like a real ID, sorta.
         staticIDCounter++;
     }
     
-    public synchronized float getCurrentBid() { return currentBid; }
+    public synchronized Double getCurrentBid() { return currentBid; }
     
     public synchronized String getCurrentHighestBidderID() { return currentHighestBidderID; }
     
@@ -40,7 +40,7 @@ public class Item
      * @param newBid
      * @param newBiddingKey Agent who is making the bid of amount newBid
      */
-    public synchronized void setCurrentBidAndBidder(float newBid, String newBiddingKey)
+    public synchronized void setCurrentBidAndBidder(Double newBid, String newBiddingKey)
     {
         currentBid = newBid;
         currentHighestBidderID = newBiddingKey;
